move proposals contianer into components folder
