+++ src/routes/governance/components/proposal-container/index.tsx
@@ -1,0 +1,1 @@
+export { ProposalContainer, PROPOSAL_QUERY } from "./proposal-container";
