+++ src/routes/governance/index.tsx
@@ -4,7 +4,7 @@ import { Route, Switch, useRouteMatch } from "react-router-dom";
 
 import { useDocumentTitle } from "../../hooks/use-document-title";
 import { RouteChildProps } from "..";
-import { ProposalContainer } from "./proposal-container";
+import { ProposalContainer } from "./components/proposal-container";
 import { ProposalsContainer } from "./proposals-container";
 
 const GovernanceRouter = ({ name }: RouteChildProps) => {
